requirements_utf8.txt
alembic==1.17.2
annotated-types==0.7.0
attrs==25.4.0
Authlib==1.6.5
beautifulsoup4==4.12.2
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
dnspython==2.8.0
filelock==3.20.0
Flask==3.1.2
Flask-Authlib-Client==0.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fpdf==1.7.2
fsspec==2025.10.0

google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.2.0
httplib2==0.31.0

humanfriendly==10.0
icalendar==6.3.2
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
lxml==5.1.0
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6

packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
gevent==25.9.1
gevent-websocket==0.10.1
pooch==1.8.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PyJWT==2.10.1
PyMatting==1.1.14
PyMySQL==1.1.0
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.3
python-socketio==5.14.3
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
sympy==1.14.0
tifffile==2025.10.16
tokenizers==0.22.1

tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.3.1
zipp==3.23.0
zope.event==6.1
zope.interface==8.1.1

groq==0.37.1
pypdf2==3.0.1
python_pptx==1.0.2⌊吠獥楴杮搠灥湥敤据敩ੳ祰整瑳㴽⸸⸳਴祰整瑳昭慬歳㴽⸱⸳ਰ祰整瑳挭癯㴽⸶⸰ਰ祰整瑳洭捯㵫㌽ㄮ⸴ਰ慆敫㵲㌽⸳⸱ਰ慦瑣牯⵹潢㵹㌽㌮ㄮ